requirements.txt.txt
streamlit
openai-whisper
indic-transliteration
requests
pydub
noisereduce
soundfile
librosa
scipy